requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-docx
pdfminer.six
python-multipart
pydantic
python-dotenv